Split tx submit and poll status in two files
