+++ crates/sargon/src/system/sargon_os/transactions/mod.rs
@@ -1,7 +1,9 @@
 mod sargon_os_transaction_analysis;
-mod sargon_os_transactions;
+mod sargon_os_transaction_status;
+mod sargon_os_transaction_submit;
 mod support;
 
 pub use sargon_os_transaction_analysis::*;
-pub use sargon_os_transactions::*;
+pub use sargon_os_transaction_status::*;
+pub use sargon_os_transaction_submit::*;
 pub use support::*;
